requirements.txt
asgiref==3.4.1
Django==3.2.9
django-cors-headers==3.10.1
django-filter==21.1
djangorestframework==3.12.4
gunicorn==20.1.0
importlib-metadata==4.8.2
Markdown==3.3.6
pytz==2021.3
Serializer==0.2.1
simplejson==3.17.6
sqlparse==0.4.2
zipp==3.6.0
